(mostly) Final Submission of Project 5/25/2017
This is the (near) final submission of my final project!

Content needs a bit of update and there’s a few css bugs I’d like to
fix. Moises helped me a lot with the flip function!

Hamburger menus are also pretty straightforward, and I used a code pen
from someone to help me template it.